remote: Split subprocess SSH transport into its own module
Extract subprocess-based SSH code (SshRemoteConnection, SshSocket,
MasterProcess, build_command_*) from ssh.rs into subprocess_ssh.rs.
ssh.rs now only contains shared types (SshConnectionOptions,
SshConnectionHost, port forward parsing). The subprocess_ssh module
is cfg-gated at the module level in transport.rs, eliminating all
per-item #[cfg(not(feature = "russh-transport"))] attributes.

Also routes SSH reconnection through RusshRemoteConnection on iOS
via feature-gated cfg in ConnectionPool::connect.
